layout & app components refactor
