+++ frontend/src/components/App/Footer/Index.tsx
@@ -1,0 +1,7 @@
+export const AppFooter: React.FC = () => {
+    return (
+        <footer className="d-flex justify-content-center flex-md-nowrap align-items-center pt-3 pb-2 border-top border-1">
+            AppFooter component
+        </footer>
+    );
+};
