--- frontend/src/components/AppFooter/Index.tsx
@@ -1,7 +1,0 @@
-export const AppFooter: React.FC = () => {
-    return (
-        <footer className="d-flex justify-content-center flex-md-nowrap align-items-center pt-3 pb-2">
-            AppFooter component
-        </footer>
-    );
-};
